+++ games_services/lib/games_services.dart
@@ -9,4 +9,5 @@ export 'src/models/leaderboard_score_data.dart';
 export 'src/models/player_data.dart';
 export 'src/models/saved_game.dart';
 export 'src/player.dart';
+export 'src/playfab/games_services_playfab.dart';
 export 'src/save_game.dart';